Pull tests out to a separate file
The tests are all written against the public API, so let's put them in
their own file, so they will be compiled separately, enforcing the API
boundary.
